+++ cli/command_create.go
@@ -94,7 +94,7 @@ func create(c *cli.Context) {
 	config.Print(releaseman.FullMode)
 
 	if !releaseman.IsCIMode {
-		ok, err := goinp.AskForBool("Are you ready for release?")
+		ok, err := goinp.AskForBoolWithDefault("Are you ready for release?", true)
 		if err != nil {
 			log.Fatalf("Failed to ask for input, error: %s", err)
 		}
@@ -124,7 +124,7 @@ func create(c *cli.Context) {
 		log.Fatalf("Failed to get latest commit, error: %#v", err)
 	}
 
-	taggedCommits, err := git.TaggedCommits()
+	taggedCommits, err := git.VersionTaggedCommits()
 	if err != nil {
 		log.Fatalf("Failed to get tagged commits, error: %#v", err)
 	}