+++ cli/command_create_release.go
@@ -88,7 +88,7 @@ func createRelease(c *cli.Context) {
 	config.Print(releaseman.ReleaseMode)
 
 	if !releaseman.IsCIMode {
-		ok, err := goinp.AskForBool("Are you ready for release?")
+		ok, err := goinp.AskForBoolWithDefault("Are you ready for release?", true)
 		if err != nil {
 			log.Fatalf("Failed to ask for input, error: %s", err)
 		}