+++ cli/command_init.go
@@ -60,7 +60,7 @@ func initRelease(c *cli.Context) {
 		if releaseman.IsCIMode {
 			log.Fatalf("Release config already exist at (%s)", releaseman.DefaultConfigPth)
 		} else {
-			ok, err := goinp.AskForBool(fmt.Sprintf("Release config already exist at (%s), would you like to overwrite it?", releaseman.DefaultConfigPth))
+			ok, err := goinp.AskForBoolWithDefault(fmt.Sprintf("Release config already exist at (%s), would you like to overwrite it?", releaseman.DefaultConfigPth), true)
 			if err != nil {
 				log.Fatalf("Failed to ask for bool, error: %#v", err)
 			} else if !ok {
@@ -73,7 +73,9 @@ func initRelease(c *cli.Context) {
 	if err != nil {
 		log.Fatalf("Failed to collect config params, error: %#v", err)
 	}
-	releaseConfig.Changelog.ItemTemplate = releaseman.ChangelogTemplate
+	releaseConfig.Changelog.ContentTemplate = releaseman.ChangelogContentTemplate
+	releaseConfig.Changelog.HeaderTemplate = releaseman.ChangelogHeaderTemplate
+	releaseConfig.Changelog.FooterTemplate = releaseman.ChangelogFooterTemplate
 
 	//
 	// Print config