+++ releaseman/changelog.go
@@ -2,6 +2,8 @@ package releaseman
 
 import (
 	"bytes"
+	"fmt"
+	"strings"
 	"text/template"
 	"time"
 
@@ -14,15 +16,23 @@ import (
 // Consts
 //=======================================
 
-// ChangelogTemplate ...
-const ChangelogTemplate = `{{range .Sections}}### {{.StartTaggedCommit.Tag}} - {{.EndTaggedCommit.Tag}} ({{.EndTaggedCommit.Date.Format "2006 Jan 02"}})
+const separator = "-----------------"
 
-{{range $idx, $commit := .Commits}}* [{{trimm $commit.Hash 7}}] {{$commit.Author}} - {{$commit.Message}} ({{$commit.Date.Format "2006 Jan 02"}})
+// ChangelogHeaderTemplate ...
+const ChangelogHeaderTemplate = `## Changelog (Current version: {{.Version}})`
+
+// ChangelogFooterTemplate ...
+const ChangelogFooterTemplate = `Updated: {{.CurrentDate.Format "2006 Jan 02"}}`
+
+// ChangelogContentTemplate ...
+const ChangelogContentTemplate = `{{range .ContentItems}}### {{.StartTaggedCommit.Tag}} - {{.EndTaggedCommit.Tag}} ({{.EndTaggedCommit.Date.Format "2006 Jan 02"}})
+
+{{range .Commits}}* [{{firstChars .Hash 7}}] {{.Author}} - {{.Message}} ({{.Date.Format "2006 Jan 02"}})
 {{end}}
 {{end}}`
 
 var changelogTemplateFuncMap = template.FuncMap{
-	"trimm": func(str string, length int) string {
+	"firstChars": func(str string, length int) string {
 		if len(str) < length {
 			return str
 		}
@@ -35,17 +45,18 @@ var changelogTemplateFuncMap = template.FuncMap{
 // Models
 //=======================================
 
-// ChangelogSectionModel ...
-type ChangelogSectionModel struct {
+// ChangelogContentItemModel ...
+type ChangelogContentItemModel struct {
 	StartTaggedCommit git.CommitModel
 	EndTaggedCommit   git.CommitModel
 	Commits           []git.CommitModel
 }
 
-// ChangelogModel ...
+// ChangelogModel ..
 type ChangelogModel struct {
-	Version  string
-	Sections []ChangelogSectionModel
+	ContentItems []ChangelogContentItemModel
+	Version      string
+	CurrentDate  time.Time
 }
 
 //=======================================
@@ -73,18 +84,19 @@ func commitsBetween(startDate *time.Time, endDate *time.Time, commits []git.Comm
 	return relevantCommits
 }
 
-func reversedSections(sections []ChangelogSectionModel) []ChangelogSectionModel {
-	reversed := []ChangelogSectionModel{}
+func reversedSections(sections []ChangelogContentItemModel) []ChangelogContentItemModel {
+	reversed := []ChangelogContentItemModel{}
 	for i := len(sections) - 1; i >= 0; i-- {
 		reversed = append(reversed, sections[i])
 	}
 	return reversed
 }
 
-func generateChangelog(commits, taggedCommits []git.CommitModel, version string) ChangelogModel {
-	changelog := ChangelogModel{
-		Version:  version,
-		Sections: []ChangelogSectionModel{},
+func generateChangelogContent(commits, taggedCommits []git.CommitModel, version string) ChangelogModel {
+	content := ChangelogModel{
+		ContentItems: []ChangelogContentItemModel{},
+		Version:      version,
+		CurrentDate:  time.Now(),
 	}
 
 	if len(taggedCommits) > 0 {
@@ -96,19 +108,19 @@ func generateChangelog(commits, taggedCommits []git.CommitModel, version string)
 
 				relevantCommits := commitsBetween(&(startTaggedCommit.Date), &(endTaggedCommit.Date), commits)
 
-				section := ChangelogSectionModel{
+				contentItem := ChangelogContentItemModel{
 					StartTaggedCommit: startTaggedCommit,
 					EndTaggedCommit:   endTaggedCommit,
 					Commits:           relevantCommits,
 				}
-				changelog.Sections = append(changelog.Sections, section)
+				content.ContentItems = append(content.ContentItems, contentItem)
 			}
 		}
 
 		// Commits between last tag and current state
 		relevantCommits := commitsBetween(&(taggedCommits[len(taggedCommits)-1].Date), nil, commits)
 
-		section := ChangelogSectionModel{
+		contentItem := ChangelogContentItemModel{
 			StartTaggedCommit: taggedCommits[len(taggedCommits)-1],
 			EndTaggedCommit: git.CommitModel{
 				Tag:  version,
@@ -116,11 +128,11 @@ func generateChangelog(commits, taggedCommits []git.CommitModel, version string)
 			},
 			Commits: relevantCommits,
 		}
-		changelog.Sections = append(changelog.Sections, section)
+		content.ContentItems = append(content.ContentItems, contentItem)
 	} else {
 		relevantCommits := commitsBetween(nil, nil, commits)
 
-		section := ChangelogSectionModel{
+		contentItem := ChangelogContentItemModel{
 			StartTaggedCommit: git.CommitModel{},
 			EndTaggedCommit: git.CommitModel{
 				Tag:  version,
@@ -128,12 +140,68 @@ func generateChangelog(commits, taggedCommits []git.CommitModel, version string)
 			},
 			Commits: relevantCommits,
 		}
-		changelog.Sections = append(changelog.Sections, section)
+		content.ContentItems = append(content.ContentItems, contentItem)
 	}
 
-	changelog.Sections = reversedSections(changelog.Sections)
+	content.ContentItems = reversedSections(content.ContentItems)
 
-	return changelog
+	return content
+}
+
+func parseChangelog(changelog string) (string, string, string) {
+	log.Debug()
+	log.Debug()
+
+	layoutLines := strings.Split(changelog, "\n")
+
+	contentStartIdx := -1
+	contentEndIdx := -1
+
+	headerStr := ""
+	for idx, line := range layoutLines {
+		if headerStr == "" {
+			headerStr = line
+		} else {
+			headerStr += fmt.Sprintf("\n%s", line)
+		}
+
+		if line == separator {
+			if headerStr != "" {
+				headerStr += "\n"
+			}
+
+			contentStartIdx = idx + 1
+			break
+		}
+	}
+
+	footerStr := ""
+	for i := (len(layoutLines) - 1); i >= 0; i-- {
+		line := layoutLines[i]
+
+		if footerStr == "" {
+			footerStr = line
+		} else {
+			footerStr = fmt.Sprintf("%s\n%s", line, footerStr)
+		}
+
+		if line == separator {
+			contentEndIdx = i - 1
+			break
+		}
+	}
+
+	contentStr := ""
+	for i := contentStartIdx; i <= contentEndIdx; i++ {
+		line := layoutLines[i]
+		if contentStr == "" {
+			contentStr = line
+		} else {
+			contentStr += fmt.Sprintf("\n%s", line)
+		}
+	}
+
+	return headerStr, footerStr, contentStr
 }
 
 //=======================================
@@ -142,36 +210,128 @@ func generateChangelog(commits, taggedCommits []git.CommitModel, version string)
 
 // WriteChangelog ...
 func WriteChangelog(commits, taggedCommits []git.CommitModel, config Config, append bool) error {
-	changelog := generateChangelog(commits, taggedCommits, config.Release.Version)
-	log.Debugf("Changelog: %#v", changelog)
+	newChangelog := generateChangelogContent(commits, taggedCommits, config.Release.Version)
 
-	changelogItemTemplateStr := ChangelogTemplate
-	if config.Changelog.ItemTemplate != "" {
-		changelogItemTemplateStr = config.Changelog.ItemTemplate
+	headerStr := ""
+	footerStr := ""
+	contentStr := ""
+
+	//
+	// Generate changelog header
+	if config.Changelog.HeaderTemplate == "" && config.Changelog.FooterTemplate == "" {
+		log.Debug()
+		log.Debug("Write changelog WITHOUT header and footer template")
 	}
 
-	tmpl := template.New("changelog").Funcs(changelogTemplateFuncMap)
-	tmpl, err := tmpl.Parse(changelogItemTemplateStr)
+	// Header
+	if config.Changelog.HeaderTemplate != "" {
+		log.Debug()
+		log.Debug("Write changelog with header and footer template")
+
+		headerTemplate := template.New("changelog_header").Funcs(changelogTemplateFuncMap)
+		headerTemplate, err := headerTemplate.Parse(ChangelogHeaderTemplate)
+		if err != nil {
+			log.Fatalf("Failed to parse header template, error: %#v", err)
+		}
+
+		var headerBytes bytes.Buffer
+		err = headerTemplate.Execute(&headerBytes, newChangelog)
+		if err != nil {
+			log.Fatalf("Failed to execute layout template, error: %#v", err)
+		}
+		headerStr = headerBytes.String()
+		headerStr += "\n\n" + separator + "\n"
+	}
+
+	// Footer
+	if config.Changelog.FooterTemplate != "" {
+		footerTemplate := template.New("changelog_footer").Funcs(changelogTemplateFuncMap)
+		footerTemplate, err := footerTemplate.Parse(ChangelogFooterTemplate)
+		if err != nil {
+			log.Fatalf("Failed to parse footer template, error: %#v", err)
+		}
+
+		var footerBytes bytes.Buffer
+		err = footerTemplate.Execute(&footerBytes, newChangelog)
+		if err != nil {
+			log.Fatalf("Failed to execute footer template, error: %#v", err)
+		}
+		footerStr = footerBytes.String()
+		footerStr = separator + "\n\n" + footerStr
+	}
+
+	log.Debug()
+	log.Debug("Layout header: %s", headerStr)
+	log.Debug("Layout footer: %s", footerStr)
+
+	//
+	// Generate changelog content
+	changelogContentTemplateStr := ChangelogContentTemplate
+	if config.Changelog.ContentTemplate != "" {
+		changelogContentTemplateStr = config.Changelog.ContentTemplate
+	}
+
+	contentTemplate := template.New("changelog_content").Funcs(changelogTemplateFuncMap)
+	contentTemplate, err := contentTemplate.Parse(changelogContentTemplateStr)
 	if err != nil {
-		log.Fatalf("Failed to parse template, error: %#v", err)
+		log.Fatalf("Failed to parse content template, error: %#v", err)
 	}
 
-	var changelogBytes bytes.Buffer
-	err = tmpl.Execute(&changelogBytes, changelog)
+	var newContentBytes bytes.Buffer
+	err = contentTemplate.Execute(&newContentBytes, newChangelog)
 	if err != nil {
 		log.Fatalf("Failed to execute template, error: %#v", err)
 	}
-	changelogStr := changelogBytes.String()
+	newContentStr := newContentBytes.String()
 
+	newContentSplit := strings.Split(newContentStr, "\n")
+	newContentSplit = newContentSplit[0 : len(newContentSplit)-1]
+	newContentStr = strings.Join(newContentSplit, "\n")
+
+	log.Debug()
+	log.Debug("Content:")
+	for _, line := range strings.Split(newContentStr, "\n") {
+		log.Debug("%s", line)
+	}
+
+	// Join header and content
 	if append {
-		log.Debugln("Changelog exist, append new version")
-		previosChangelogStr, err := fileutil.ReadStringFromFile(config.Changelog.Path)
+		log.Debug()
+		log.Debug("Previous changelog exist, append new conent")
+
+		prevChangelogStr, err := fileutil.ReadStringFromFile(config.Changelog.Path)
 		if err != nil {
 			return err
 		}
 
-		changelogStr = changelogStr + previosChangelogStr
+		prevContentStr := ""
+		if config.Changelog.HeaderTemplate != "" && config.Changelog.FooterTemplate != "" {
+			_, _, prevContentStr = parseChangelog(prevChangelogStr)
+		} else {
+			prevContentStr = prevChangelogStr
+		}
+
+		log.Debug()
+		log.Debug("Prev content:")
+		for _, line := range strings.Split(prevContentStr, "\n") {
+			log.Debug("%s", line)
+		}
+
+		contentStr = fmt.Sprintf("%s\n%s", newContentStr, prevContentStr)
+
+		log.Debug()
+		log.Debug("Merged content:")
+		for _, line := range strings.Split(contentStr, "\n") {
+			log.Debug("%s", line)
+		}
+	} else {
+		log.Debug()
+		log.Debug("NO previous changelog exist")
+
+		contentStr = newContentStr
 	}
+
+	changelogStr := headerStr + "\n" + contentStr + "\n" + footerStr
 
 	return fileutil.WriteStringToFile(config.Changelog.Path, changelogStr)
 }